# docs/swagger.yaml
definitions:
  models.Actor:
    properties:
      birth_date:
        type: string
      created_at:
        type: string
      deleted_at:
        type: string
      gender:
        type: string
      id:
        type: string
      name:
        minLength: 3
        type: string
      updated_at:
        type: string
    required:
    - name
    type: object
  models.ActorSwagger:
    properties:
      birth_date:
        type: string
      gender:
        type: string
      name:
        minLength: 3
        type: string
    required:
    - name
    type: object
  models.ActorsListResp:
    properties:
      actors:
        items:
          $ref: '#/definitions/models.Actor'
        type: array
      count:
        type: integer
    type: object
  models.Film:
    properties:
      created_at:
        type: string
      deleted_at:
        type: string
      description:
        type: string
      film_actors:
        items:
          $ref: '#/definitions/models.FilmActor'
        type: array
      id:
        type: string
      rating:
        type: integer
      release_date:
        type: string
      title:
        type: string
      updated_at:
        type: string
    type: object
  models.FilmActor:
    properties:
      actor:
        $ref: '#/definitions/models.Actor'
      actor_id:
        type: string
      created_at:
        type: string
      deleted_at:
        type: string
      film_id:
        type: string
      id:
        type: string
    type: object
  models.FilmSwagger:
    properties:
      actor_ids:
        items:
          type: string
        type: array
      description:
        type: string
      rating:
        type: integer
      release_date:
        type: string
      title:
        type: string
    type: object
  models.FilmsListResp:
    properties:
      count:
        type: integer
      films:
        items:
          $ref: '#/definitions/models.Film'
        type: array
    type: object
info:
  contact: {}
paths:
  /actors:
    post:
      consumes:
      - application/json
      description: create new actor
      parameters:
      - description: body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.ActorSwagger'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Actor'
        "500":
          description: Internal Server Error
          schema: {}
      summary: create new actor
      tags:
      - Actor
  /actors/{id}:
    delete:
      consumes:
      - application/json
      description: delete actor
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema: {}
      summary: Delete actor
      tags:
      - Actor
    get:
      consumes:
      - application/json
      description: Get actor by id
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Actor'
        "500":
          description: Internal Server Error
          schema: {}
      summary: Get actor
      tags:
      - Actor
    put:
      consumes:
      - application/json
      description: update new actor
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      - description: body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.ActorSwagger'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ActorSwagger'
        "500":
          description: Internal Server Error
          schema: {}
      summary: Update actor
      tags:
      - Actor
  /actors/list:
    get:
      consumes:
      - application/json
      description: Get all actor
      parameters:
      - description: search by name
        in: query
        name: search
        type: string
      - description: page number
        format: page
        in: query
        name: page
        type: integer
      - description: number of elements per page
        format: limit
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ActorsListResp'
        "500":
          description: Internal Server Error
          schema: {}
      summary: Get Actor
      tags:
      - Actor
  /films:
    post:
      consumes:
      - application/json
      description: create new film
      parameters:
      - description: body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.FilmSwagger'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Film'
        "500":
          description: Internal Server Error
          schema: {}
      summary: create new film
      tags:
      - Film
  /films/list:
    get:
      consumes:
      - application/json
      description: Get all films
      parameters:
      - description: search by title
        in: query
        name: search
        type: string
      - description: page number
        format: page
        in: query
        name: page
        type: integer
      - description: number of elements per page
        format: limit
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.FilmsListResp'
        "500":
          description: Internal Server Error
          schema: {}
      summary: Get Films
      tags:
      - Film
swagger: "2.0"